Add a setup.py install file
