+++ idarling_server.py
@@ -10,104 +10,8 @@
 
 # You should have received a copy of the GNU General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
-import argparse
-import logging
-import os
-import signal
-import sys
-import traceback
-
-from PyQt5.QtCore import QCoreApplication, QTimer
-
-from idarling.shared.server import Server
+from idarling import server
 
 
-class DedicatedServer(Server):
-    """
-    The dedicated server implementation.
-    """
-
-    def __init__(self, ssl, parent=None):
-        logger = self.start_logging()
-        Server.__init__(self, logger, ssl, parent)
-
-    def local_file(self, filename):
-        filesDir = os.path.join(os.path.dirname(__file__), 'files')
-        filesDir = os.path.abspath(filesDir)
-        if not os.path.exists(filesDir):
-            os.makedirs(filesDir)
-        return os.path.join(filesDir, filename)
-
-    def start_logging(self):
-        logger = logging.getLogger('IDArling.Server')
-
-        # Get path to the log file
-        logDir = os.path.join(os.path.dirname(__file__), 'logs')
-        logDir = os.path.abspath(logDir)
-        if not os.path.exists(logDir):
-            os.makedirs(logDir)
-        logPath = os.path.join(logDir, 'idarling.%s.log' % os.getpid())
-
-        # Configure the logger
-        logger.setLevel(logging.DEBUG)
-        logFormat = '[%(asctime)s][%(levelname)s] %(message)s'
-        formatter = logging.Formatter(fmt=logFormat, datefmt='%H:%M:%S')
-
-        # Log to the console
-        streamHandler = logging.StreamHandler()
-        streamHandler.setFormatter(formatter)
-        logger.addHandler(streamHandler)
-
-        # Log to the log file
-        fileHandler = logging.FileHandler(logPath)
-        fileHandler.setFormatter(formatter)
-        logger.addHandler(fileHandler)
-
-        return logger
-
-
-def main(args):
-    """
-    The entry point of a Python program.
-    """
-    app = QCoreApplication(sys.argv)
-    sys.excepthook = traceback.print_exception
-
-    server = DedicatedServer(args.ssl)
-    server.start(args.host, args.port)
-
-    # Allow the use of Ctrl-C to stop the server
-    def sigint_handler(signum, frame):
-        server.stop()
-        app.exit(0)
-    signal.signal(signal.SIGINT, sigint_handler)
-
-    def safe_timer(timeout, func, *args, **kwargs):
-        def timer_event():
-            try:
-                func(*args, **kwargs)
-            finally:
-                QTimer.singleShot(timeout, timer_event)
-        QTimer.singleShot(timeout, timer_event)
-    safe_timer(50, lambda: None)
-    return app.exec_()
-
-
-if __name__ == '__main__':
-    parser = argparse.ArgumentParser(add_help=False)
-    parser.add_argument('--help', action='help',
-                        help='show this help message and exit')
-
-    parser.add_argument('-h', '--host', type=str, default='127.0.0.1',
-                        help='the hostname to start listening on')
-    parser.add_argument('-p', '--port', type=int, default=31013,
-                        help='the port to start listening on')
-
-    security = parser.add_mutually_exclusive_group(required=True)
-    security.add_argument('--ssl', type=str, nargs=2,
-                          metavar=('fullchain.pem', 'privkey.pem'),
-                          help='the certificate and private key files')
-    security.add_argument('--no-ssl', action='store_true',
-                          help='disable SSL (not recommended)')
-
-    main(parser.parse_args())
+if __name__ == "__main__":
+    server.main()
